Restructured directories. Working on origin heats parser.
