+++ python/attic/oldcode/Constants.py
