+++ python/attic/oldcode/ITC-sampl4.py
