+++ python/attic/oldcode/pymc-classmodel.py
