+++ python/attic/oldcode/pymc-test.py
